# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.EntitySchemaRequest:
    properties:
      entity_type:
        example: user
        type: string
      indexes:
        items:
          $ref: '#/definitions/models.IndexConfig'
        type: array
      properties:
        $ref: '#/definitions/models.PropertySchema'
    required:
    - entity_type
    - properties
    type: object
  handlers.EntitySchemaResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      entity_type:
        example: user
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      indexes:
        items:
          $ref: '#/definitions/models.IndexConfig'
        type: array
      properties:
        $ref: '#/definitions/models.PropertySchema'
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      version:
        example: 1
        type: integer
    type: object
...
  handlers.RelationshipSchemaRequest:
    properties:
      cardinality:
        allOf:
        - $ref: '#/definitions/models.CardinalityType'
        enum:
        - one-to-one
        - one-to-many
        - many-to-one
        - many-to-many
      denormalization_config:
        $ref: '#/definitions/models.DenormalizationConfig'
      from_entity_type:
        example: user
        type: string
      properties:
        $ref: '#/definitions/models.PropertySchema'
      relationship_type:
        example: owns
        type: string
      to_entity_type:
        example: document
        type: string
    required:
    - cardinality
    - from_entity_type
    - relationship_type
    - to_entity_type
    type: object
  handlers.RelationshipSchemaResponse:
    properties:
      cardinality:
        allOf:
        - $ref: '#/definitions/models.CardinalityType'
        enum:
        - one-to-one
        - one-to-many
        - many-to-one
        - many-to-many
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      denormalization_config:
        $ref: '#/definitions/models.DenormalizationConfig'
      from_entity_type:
        example: user
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      properties:
        $ref: '#/definitions/models.PropertySchema'
      relationship_type:
        example: owns
        type: string
      to_entity_type:
        example: document
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      version:
        example: 1
        type: integer
    type: object
  middleware.ErrorDetail:
    description: Detailed error information including code, message, and optional
      details
    properties:
      code:
        example: NOT_FOUND
        type: string
      details:
        type: object
      message:
        example: Entity not found
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  middleware.ErrorResponse:
    description: Standard error response format for all API errors
    properties:
      error:
        $ref: '#/definitions/middleware.ErrorDetail'
    type: object
  models.CardinalityType:
    enum:
    - one-to-one
    - one-to-many
    - many-to-one
    - many-to-many
    type: string
    x-enum-varnames:
    - OneToOne
    - OneToMany
    - ManyToOne
    - ManyToMany
  models.DenormalizationConfig:
    properties:
      denormalize_from_to:
        example:
        - title
        - status
        items:
          type: string
        type: array
      denormalize_to_from:
        example:
        - name
        - email
        items:
          type: string
        type: array
      include_relation_data:
        example: false
        type: boolean
      update_on_change:
        example: true
        type: boolean
    type: object
  models.IndexConfig:
    properties:
      fields:
        example:
        - email
        items:
          type: string
        minItems: 1
        type: array
      name:
        example: idx_user_email
        type: string
      type:
        enum:
        - btree
        - hash
        - gin
        - gist
        - vector
        example: btree
        type: string
      unique:
        example: true
        type: boolean
      vector_type:
        example: ivfflat
        type: string
    required:
    - fields
    - name
    - type
    type: object
  models.PropertyDefinition:
    properties:
      constraints:
        type: object
      default:
        type: object
      description:
        example: User's full name
        type: string
      element_type:
        example: string
        type: string
      object_schema:
        $ref: '#/definitions/models.PropertySchema'
      required:
        example: true
        type: boolean
      type:
        enum:
        - string
        - number
        - boolean
        - datetime
        - object
        - array
        - vector
        example: string
        type: string
      vector_dim:
        example: 512
        type: integer
    required:
    - type
    type: object
  models.PropertySchema:
    additionalProperties:
      $ref: '#/definitions/models.PropertyDefinition'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@entropic.io
    name: API Support
    url: http://www.entropic.io/support
  description: A next-generation storage engine with flexible entity-relationship
    model and dual-storage architecture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Entropic Storage Engine API
  version: "1.0"
paths:
...
  /api/v1/schemas/entities:
    get:
      consumes:
      - application/json
      description: Get a list of all entity schemas in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.EntitySchemaResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: List all entity schemas
      tags:
      - schemas
    post:
      consumes:
      - application/json
      description: Create a new schema for an entity type
      parameters:
      - description: Entity schema data
        in: body
        name: schema
        required: true
        schema:
          $ref: '#/definitions/handlers.EntitySchemaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.EntitySchemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Create an entity schema
      tags:
      - schemas
  /api/v1/schemas/entities/{entityType}:
    delete:
      consumes:
      - application/json
      description: Delete the schema for an entity type. This operation cannot be
        undone.
      parameters:
      - description: Entity Type
        in: path
        name: entityType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Delete an entity schema
      tags:
      - schemas
    get:
      consumes:
      - application/json
      description: Get the schema for a specific entity type
      parameters:
      - description: Entity Type
        in: path
        name: entityType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EntitySchemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Get an entity schema
      tags:
      - schemas
    put:
      consumes:
      - application/json
      description: Update the schema for an existing entity type
      parameters:
      - description: Entity Type
        in: path
        name: entityType
        required: true
        type: string
      - description: Entity schema update data
        in: body
        name: schema
        required: true
        schema:
          $ref: '#/definitions/handlers.EntitySchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EntitySchemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Update an entity schema
      tags:
      - schemas
  /api/v1/schemas/relationships:
    get:
      consumes:
      - application/json
      description: Get a list of all relationship schemas in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.RelationshipSchemaResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: List all relationship schemas
      tags:
      - schemas
    post:
      consumes:
      - application/json
      description: Create a new schema for a relationship type
      parameters:
      - description: Relationship schema data
        in: body
        name: schema
        required: true
        schema:
          $ref: '#/definitions/handlers.RelationshipSchemaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RelationshipSchemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Create a relationship schema
      tags:
      - schemas
  /api/v1/schemas/relationships/{relationshipType}:
    delete:
      consumes:
      - application/json
      description: Delete the schema for a relationship type. This operation cannot
        be undone.
      parameters:
      - description: Relationship Type
        in: path
        name: relationshipType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Delete a relationship schema
      tags:
      - schemas
    get:
      consumes:
      - application/json
      description: Get the schema for a specific relationship type
      parameters:
      - description: Relationship Type
        in: path
        name: relationshipType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RelationshipSchemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Get a relationship schema
      tags:
      - schemas
    put:
      consumes:
      - application/json
      description: Update the schema for an existing relationship type
      parameters:
      - description: Relationship Type
        in: path
        name: relationshipType
        required: true
        type: string
      - description: Relationship schema update data
        in: body
        name: schema
        required: true
        schema:
          $ref: '#/definitions/handlers.RelationshipSchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RelationshipSchemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Update a relationship schema
      tags:
      - schemas
...
schemes:
- http
- https
swagger: "2.0"